refactored everything in RobotApp
